# DoH/DoT/TCP-based lookups and connection re-use