--- a/src/cli/worker_manager.go
+++ b/src/cli/worker_manager.go
@@ -458,7 +458,6 @@ type InputLineWithNameServer struct {
 // The goal is that a query for a single name server will consistently go to the same worker pool which 1+ threads will read from
 // This is especially useful for TLS/TCP/HTTPS based lookups where repeating the initial handshakes would be wasteful
 func inputDeMultiplexer(nameServers []zdns.NameServer, inChan <-chan string, workerPools *WorkerPools, wg *sync.WaitGroup) error {
-	wg.Add(1)
 	defer wg.Done()
 	// defer closing the worker pool chans
 	defer func() {
@@ -553,12 +552,16 @@ func Run(gc CLIConf) {
 			log.Fatal(fmt.Sprintf("could not feed input channel: %v", inErr))
 		}
 	}()
-	go func() {
-		plexErr := inputDeMultiplexer(uniqNameServers, inChan, workerPools, &routineWG)
-		if plexErr != nil {
-			log.Fatal(fmt.Sprintf("could not de-multiplex input channel: %v", plexErr))
-		}
-	}()
+	const numberOfDeMultiplexers = 5
+	for i := 0; i < numberOfDeMultiplexers; i++ {
+		go func() {
+			plexErr := inputDeMultiplexer(uniqNameServers, inChan, workerPools, &routineWG)
+			if plexErr != nil {
+				log.Fatal(fmt.Sprintf("could not de-multiplex input channel: %v", plexErr))
+			}
+		}()
+	}
+	routineWG.Add(numberOfDeMultiplexers)
 	go func() {
 		outErr := outHandler.WriteResults(outChan, &routineWG)
 		if outErr != nil {