@phillip-stephens 2024-09-13
After talking with Zakir, we can make this quite a bit simpler if we just re-use the existing connection in `ExternalLookup` if the user/CLI doesn't suggest a new one. Closing this but leaving branch in case we ever want to re-visit. #445 has the new approach